Rename and move MusicMixer to Creator/MusicCreator
