+++ Assets/GothicVR/Scripts/Creator/MusicCreator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9a2b6d5503112494f89b3a4c775c37d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
